Modernize package and CI (#69)
While starting to review #68, I felt the pain of CI / package metadata
not being set up to modern norms. Here's a few things this PR does, and
I'll make a note if I think we should consider doing this as a follow
up.

Question: **This does not, and should not, change anything about python
functionality. All python code touched is from ruff format and check,
which I suppose we _could_ do as a follow-up, especially because it
might be annoying to merge into #68 -- thoughts?**

1. drops setup.cfg, moves package data entirely to pyproject. 
2. modernizes pyproject, including new license format, newer pythons,
dependency-groups,
3. moves to ruff (see above question) -- and thus updates to modern
typing
4. updates pre-commit to be identical to what is used in
napari-template, using ruff
5. adds dependabot
6. updates CI (and tox) for modern python, dropping npe2 validation,
adding uv setup, and adding hynek/build-and-inspect, and using modern
deploy actions

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>